split mySiteScreen to two screens
